refactor: refine filterable skills page - Add category filter - Refine filtering on skills page - Footer style update
