+++ src/pages/skills.js
@@ -1,120 +1,225 @@
 import PropTypes from "prop-types";
 import { graphql } from "gatsby";
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { css } from "@emotion/react";
 import MainLayout from "../layouts/MainLayout";
 import Tile from "../components/Tile";
-import ProgressBar from "../components/ProgressBar";
 import Icon from "../components/Icons";
-import { SKILL_CATEGORY, siteOptions } from "../utils/constants";
+import { SKILL_CATEGORY, SKILL_RATINGS, siteOptions } from "../utils/constants";
 import PageTitle from "../components/PageTitle";
 import SEO from "../components/SEO";
 
-const SkillsPage = ({ data }) => {
-  const { nodes: skills } = data.allSkills;
+// import { navigate } from "@reach/router";
+
+const SKILL_RATINGS_BAR_COLOR = {
+  1: { light: `var(--color-red-200)`, dark: `var(--color-red-600)` },
+  2: { light: `var(--color-red-200)`, dark: `var(--color-red-600)` },
+  3: { light: `var(--color-yellow-200)`, dark: `var(--color-yellow-600)` },
+  4: { light: `var(--color-yellow-200)`, dark: `var(--color-yellow-600)` },
+  5: { light: `var(--color-green-200)`, dark: `var(--color-green-600)` },
+};
+
+const filterByRating = (minRating) => (skill) => {
+  return minRating <= skill.rating;
+};
+
+const filterByCategory = (category) => (skill) => {
+  if (!category || category === "all") {
+    return true;
+  }
+  return SKILL_CATEGORY[skill.name] === category;
+};
+
+const SkillsPageFilterable = ({ data }) => {
+  const allSkills = data.allSkills.nodes;
+  const [skills, setSkills] = useState(allSkills);
+  const params = new URLSearchParams(location.search);
+  const minSkillLevel = params.get("minSkillLevel");
+  const category = params.get("category");
+
+  const [formState, setFormState] = useState({
+    minRating: minSkillLevel || 1,
+    category: category || "",
+  });
+
+  useEffect(() => {
+    const { minRating, category } = formState;
+    let filteredSkills = allSkills
+      .filter(filterByRating(minRating))
+      .filter(filterByCategory(category));
+
+    setSkills(filteredSkills);
+  }, [allSkills, formState]);
+
   return (
     <>
       <SEO title="Skills" />
       <MainLayout>
         <PageTitle>Skills</PageTitle>
-
         <div
           css={css`
             display: flex;
+            flex-direction: row;
+            margin: 2em 0;
+
             > * {
-              width: 50%;
+              margin: 1em 0;
+            }
+
+            > :not(:first-child) {
+              margin-left: 2em;
             }
             @media screen and (max-width: ${siteOptions.layout
                 .mobileBreakpoint}) {
-              flex-wrap: wrap;
-              > * {
-                width: 100%;
+              flex-direction: column;
+
+              > :not(:first-child) {
+                margin-left: 0;
               }
             }
           `}
         >
-          <Tile
-            icon="window"
-            css={css`
-              margin: 1em;
-            `}
-          >
-            <h2
+          <div css={css``}>
+            <label
+              htmlFor="skillFilter"
               css={css`
-                text-align: center;
+                font-weight: bold;
               `}
             >
-              Frontend
-            </h2>
-            {skills
-              .filter((skill) => SKILL_CATEGORY[skill.name] === "frontend")
-              .sort((a, b) => b.rating - a.rating)
-              .map((skill) => renderSkill(skill.name, skill.rating))}
-          </Tile>
-          <Tile
-            icon="terminal"
-            css={css`
-              margin: 1em;
-            `}
-          >
-            <h2
+              Skill level
+            </label>
+            <select
+              id="skillFilter"
               css={css`
-                text-align: center;
+                background-color: transparent;
+                margin: 0.25em 0;
+                width: 100%;
+                cursor: pointer;
+                border-radius: 0.25rem;
+                border: 1px solid var(--color-neutrals-500);
+                padding: 0.5em;
+
+                .dark-mode & {
+                  color: var(--color-dark-700);
+                  background-color: var(--color-dark-300);
+                  border: var(--color-dark-300);
+                }
+              `}
+              // onChange={(e) => handleRatingFilterSelect(e)}
+              onChange={(e) =>
+                setFormState((state) => ({
+                  ...state,
+                  minRating: e.target.value,
+                }))
+              }
+            >
+              <option value={1}>All the things</option>
+              <option value={4}>Stuff I&apos;m pretty good at</option>
+            </select>
+          </div>
+          <div css={css``}>
+            <label
+              htmlFor="categoryFilter"
+              css={css`
+                font-weight: bold;
               `}
             >
-              Backend
-            </h2>
-            {skills
-              .filter((skill) => SKILL_CATEGORY[skill.name] === "backend")
+              Category
+            </label>
+            <select
+              id="categoryFilter"
+              css={css`
+                background-color: transparent;
+                margin: 0.25em 0;
+                width: 100%;
+                cursor: pointer;
+                border-radius: 0.25rem;
+                border: 1px solid var(--color-neutrals-500);
+                padding: 0.5em;
+
+                .dark-mode & {
+                  color: var(--color-dark-700);
+                  background-color: var(--color-dark-300);
+                  border: var(--color-dark-300);
+                }
+              `}
+              // onChange={(e) => handleCategoryFilter(e)}
+              onChange={(e) =>
+                setFormState((state) => ({
+                  ...state,
+                  category: e.target.value,
+                }))
+              }
+            >
+              <option value={"all"}>All the things</option>
+              <option value={"frontend"}>Front end</option>
+              <option value={"backend"}>Back end</option>
+            </select>
+          </div>
+        </div>
+        <div
+          css={css`
+            display: grid;
+            grid-template-columns: repeat(auto-fill, minmax(120px, 1fr));
+            grid-gap: 1rem;
+          `}
+        >
+          {skills &&
+            skills
               .sort((a, b) => b.rating - a.rating)
-              .map((skill) => renderSkill(skill.name, skill.rating))}
-          </Tile>
+              .map((skill) => (
+                <Tile
+                  key={skill.name}
+                  css={css`
+                    display: flex;
+                    flex-direction: column;
+                    justify-content: center;
+                    align-items: center;
+                    text-align: center;
+                    margin: 0;
+                    padding: 0.5em;
+
+                    & :after {
+                      content: "";
+                      height: 3px;
+                      width: ${(skill.rating / SKILL_RATINGS.MAX) * 100}%;
+                      background-color: ${SKILL_RATINGS_BAR_COLOR[skill.rating]
+                        .dark};
+                      position: absolute;
+                      bottom: 0;
+                      left: 0;
+
+                      .light-mode & {
+                        background-color: ${SKILL_RATINGS_BAR_COLOR[
+                          skill.rating
+                        ].light};
+                      }
+                    }
+                  `}
+                >
+                  <div
+                    css={css`
+                      margin-bottom: 0.5em;
+                    `}
+                  >
+                    <Icon
+                      name={skill.name}
+                      viewbox="0 0 128 128"
+                      size="2.5em"
+                    />
+                  </div>
+                  <div>{skill.name}</div>
+                </Tile>
+              ))}
         </div>
       </MainLayout>
     </>
   );
 };
 
-SkillsPage.propTypes = {
+SkillsPageFilterable.propTypes = {
   data: PropTypes.object.isRequired,
 };
-
-const renderSkill = (skillName, skillRating) => (
-  <div
-    key={skillName}
-    css={css`
-      margin-bottom: 2em;
-      display: flex;
-      justify-content: space-between;
-    `}
-  >
-    <div
-      css={css`
-        text-align: center;
-        width: 5rem;
-      `}
-    >
-      <div>
-        <Icon name={skillName} viewbox="0 0 128 128" size="2em" />
-      </div>
-      <div
-        css={css`
-          margin: 0;
-          font-size: 0.8em;
-        `}
-      >
-        {skillName}
-      </div>
-    </div>
-    <div
-      css={css`
-        width: 70%;
-      `}
-    >
-      <ProgressBar currentLevel={skillRating} />
-    </div>
-  </div>
-);
 
 export const pageQuery = graphql`
   query {
@@ -129,4 +234,4 @@ export const pageQuery = graphql`
   }
 `;
 
-export default SkillsPage;
+export default SkillsPageFilterable;
